requirements.txt
fastmcp>=2.12.0
uvicorn>=0.38.0
websockets>=15.0.0
google-api-python-client>=2.149.0
google-auth>=2.35.0
google-auth-oauthlib>=1.2.1
python-dateutil>=2.9.0.post0
python-dotenv>=1.0.0
